rename tera-tab-content to tera-project-page
